I completed the majority of Allegheny County Airport Spider. It gets correct location and time information from the designated website, passed all linters except yapf (I couldn't get yapf to run on my machine for some reason), and passes rudimentary tests.
GOING FORWARD: I need to ask what meeting status and meeting id are supposed to be in the next meeting and update that bit of functionality in alle_airport.py. I also need to discuss possibilities for further testing, and should check with an experienced member of the group if the spider works as desired.