Remove old documentation, remove contextual documentation, and add new troubleshooting documentation
With the exception of `03-team.md`, the new documentation is completed
and we are prepared to remove the old documentation. At the same time,
we decided to remove the "contextual" documentation as well, since it
did not contribute directly to the needs of volunteers and would add
debt as it diverged from the original Chicago documentation.